+++ components/forms/poll-options-form/month-calendar/time-picker.tsx
@@ -1,9 +1,10 @@
-import { Combobox } from "@headlessui/react";
+import { Listbox } from "@headlessui/react";
 import clsx from "clsx";
-import { addMinutes, format, isSameDay, setHours, setMinutes } from "date-fns";
+import { addMinutes, isSameDay, setHours, setMinutes } from "date-fns";
 import * as React from "react";
 import ReactDOM from "react-dom";
 import { usePopper } from "react-popper";
+import { localeFormat } from "utils/date-time-utils";
 import { stopPropagation } from "utils/stop-propagation";
 
 import ChevronDown from "../../../icons/chevron-down.svg";
@@ -38,7 +39,6 @@ const TimePicker: React.VoidFunctionComponent<TimePickerProps> = ({
     ],
   });
 
-  const [query, setQuery] = React.useState("");
   const options: React.ReactNode[] = [];
   for (let i = 0; i < 96; i++) {
     const optionValue = addMinutes(startFrom, i * 15);
@@ -47,62 +47,48 @@ const TimePicker: React.VoidFunctionComponent<TimePickerProps> = ({
       // because react-big-calendar does not support events that span days
       break;
     }
-    if (query && !format(optionValue, "hhmma").includes(query)) {
-      continue;
-    }
     options.push(
-      <Combobox.Option
+      <Listbox.Option
         key={i}
         className={styleMenuItem}
         value={optionValue.toISOString()}
       >
-        {format(optionValue, "p")}
-      </Combobox.Option>,
+        {localeFormat(optionValue, "p")}
+      </Listbox.Option>,
     );
   }
 
   const portal = document.getElementById("portal");
 
   return (
-    <Combobox
+    <Listbox
       value={value.toISOString()}
       onChange={(newValue) => {
-        setQuery("");
         onChange?.(new Date(newValue));
       }}
     >
       <div ref={setReferenceElement} className={clsx("relative", className)}>
-        {/* Remove generic params once Combobox.Input can infer the types */}
-        <Combobox.Input<"input">
-          className="input w-28 pr-8"
-          displayValue={() => ""}
-          onChange={(e) => {
-            setQuery(e.target.value.toUpperCase().replace(/[\:\s]/g, ""));
-          }}
-        />
-        <Combobox.Button className="absolute inset-0 flex h-9 cursor-default items-center px-2 text-left">
-          <span className="grow truncate">
-            {!query ? format(value, "p") : null}
-          </span>
-          <span className="pointer-events-none flex">
+        <Listbox.Button className="btn-default text-left">
+          <span className="grow truncate">{localeFormat(value, "p")}</span>
+          <span className="pointer-events-none ml-2 flex">
             <ChevronDown className="h-5 w-5" />
           </span>
-        </Combobox.Button>
+        </Listbox.Button>
         {portal &&
           ReactDOM.createPortal(
-            <Combobox.Options
+            <Listbox.Options
               style={styles.popper}
               {...attributes.popper}
               ref={setPopperElement}
-              className="z-50 max-h-72 w-32 overflow-auto rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none"
+              className="z-50 max-h-72 overflow-auto rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none"
               onMouseDown={stopPropagation}
             >
               {options}
-            </Combobox.Options>,
+            </Listbox.Options>,
             portal,
           )}
       </div>
-    </Combobox>
+    </Listbox>
   );
 };
